telegram-ai-bot/requirements.txt
python-telegram-bot==21.9
groq==0.13.0
python-dotenv==1.0.0
